requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
Django==5.2
fastapi==0.115.12
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
